# swagger/swagger.yml
swagger: "2.0"

info:
  version: "1.0.0"
  title: "Users API"
  description: "A users API to register, login and get users."

schemes:
  - "http"
host: "localhost:3000"
basePath: "/users"

produces:
  - "text/html"
  - "application/json"

paths:
  /register:
    post:
      summary: Registers a user.
      description: Adds a new user in the database.
      parameters:
        - name: params
          in: body
          description: The user to register.
          schema:
            required:
              - email
              - password
            properties:
              email:
                type: string
              password:
                type: string
      responses:
        200:
          description: User successfully registered.
          schema:
            required:
              - status
              - token
            properties:
              status:
                type: string
              token:
                type: string
        500:
          description: User couldn't have be registed.
  /login:
    post:
      summary: Logs in a user.
      description: Logs in an existing user.
      parameters:
        - name: params
          in: body
          description: The user to log in.
          schema:
            required:
              - email
              - password
            properties:
              email:
                type: string
              password:
                type: string
      responses:
        200:
          description: User successfully logged in.
          schema:
            required:
              - status
              - token
            properties:
              status:
                type: string
              token:
                type: string
        500:
          description: User couldn't log in.
  /user/{id}:
    get:
      summary: Gets a user.
      description: Gets an existing user.
      parameters:
        - name: id
          in: path
          required: true
          description: User ID.
          type: integer
        - name: authorization
          in: header
          required: true
          description: JWT authorisation token.
          type: string
      responses:
        200:
          description: User returned.
          schema:
            required:
              - status
              - token
            properties:
              status:
                type: string
              user:
                type: integer
        400:
          description: Please log in.
          schema:
            required:
              - status
            properties:
              status:
                type: string
    put:
      summary: Updates a user.
      description: Updates an existing user.
      parameters:
        - name: id
          in: path
          required: true
          description: User ID.
          type: integer
        - name: authorization
          in: header
          required: true
          description: JWT authorisation token.
          type: string
        - name: params
          in: body
          description: The user to register.
          schema:
            required:
              - firstName
              - lastName
            properties:
              firstName:
                type: string
              lastName:
                type: string
      responses:
        201:
          description: User updated.
          schema:
            required:
              - status
            properties:
              status:
                type: string
    delete:
      summary: Delete a user.
      description: Deletes an existing user.
      parameters:
        - name: id
          in: path
          required: true
          description: User ID.
          type: integer
        - name: authorization
          in: header
          required: true
          description: JWT authorisation token.
          type: string
      responses:
        200:
          description: User deleted.
          schema:
            required:
              - status
              - user
            properties:
              status:
                type: string
              user:
                type: integer
        404:
          description: User could not be found.
          schema:
            required:
              - status
            properties:
              status:
                type: string